Solve for line numbers in includes
Also reorganize the macros into a utils lib, since that just makes more sense.
