+++ src/graph/metadata.rs
@@ -1,16 +1,16 @@
 use std::{collections::HashMap, fmt::Debug};
 
-use anyhow::{Result, anyhow};
 use once_cell::sync::Lazy;
 use regex::Regex;
 use serde::Serialize;
 
 use crate::graph::nodes::Location;
 use crate::scanner::tokens::{Token, TokenType};
+use crate::utils::{extract_attributes, key_values_from_named_attribute};
 
 // just make this quoted, and then pull everything else out
-static RE_NAMED_QUOTED: Lazy<Regex> = Lazy::new(|| Regex::new(r#"(\w*=".*?")"#).unwrap());
-static RE_NAMED: Lazy<Regex> = Lazy::new(|| Regex::new(r#"(.*?)[=|,](.*)"#).unwrap());
+pub static RE_NAMED_QUOTED: Lazy<Regex> = Lazy::new(|| Regex::new(r#"(\w*=".*?")"#).unwrap());
+pub static RE_NAMED: Lazy<Regex> = Lazy::new(|| Regex::new(r#"(.*?)[=|,](.*)"#).unwrap());
 
 #[derive(PartialEq, Clone, Debug)]
 pub enum AttributeType {
@@ -139,17 +139,7 @@ impl ElementMetadata {
             new_block_metadata.attributes.insert("id".to_string(), id);
         } else {
             let attribute_list = token.lexeme[1..token.lexeme.len() - 1].to_string();
-            let mut non_quoted_key_values = attribute_list.clone();
-            let mut attributes: Vec<&str> = vec![];
-            // TK "1,2,4" should be a single attribute, not "1,", 2, 4"
-            for quoted_attr in RE_NAMED_QUOTED.captures_iter(&attribute_list) {
-                let (total, [_]) = quoted_attr.extract();
-                attributes.push(total);
-                non_quoted_key_values = non_quoted_key_values.replace(total, "");
-            }
-            attributes.extend(non_quoted_key_values.split(',').collect::<Vec<&str>>());
-
-            new_block_metadata.process_attributes(attributes);
+            new_block_metadata.process_attributes(extract_attributes(&attribute_list));
         }
         new_block_metadata
     }
@@ -162,8 +152,7 @@ impl ElementMetadata {
             }
         } else {
             let attribute_list = token.lexeme[1..token.lexeme.len() - 1].to_string();
-            let attributes: Vec<&str> = attribute_list.split(',').collect();
-            self.process_attributes(attributes);
+            self.process_attributes(extract_attributes(&attribute_list));
         }
     }
 
@@ -181,9 +170,8 @@ impl ElementMetadata {
         self.roles.extend(incoming.options.clone());
     }
 
-    pub fn process_attributes(&mut self, mut attributes: Vec<&str>) {
+    pub fn process_attributes(&mut self, mut attributes: Vec<String>) {
         for (idx, attribute) in attributes.iter_mut().enumerate() {
-            dbg!(&attribute);
             match key_values_from_named_attribute(attribute) {
                 Ok((key, values)) => {
                     if key == *"role" {
@@ -211,7 +199,7 @@ impl ElementMetadata {
         }
     }
 
-    fn process_attribute(&mut self, idx: usize, attribute: &mut &str) {
+    fn process_attribute(&mut self, idx: usize, attribute: &mut String) {
         match self.declared_type {
             Some(AttributeType::Source) => {
                 if idx == 1 {
@@ -236,7 +224,7 @@ impl ElementMetadata {
             }
             _ => {
                 if attribute.starts_with('"') {
-                    *attribute = &attribute[1..attribute.len() - 1]
+                    *attribute = attribute[1..attribute.len() - 1].to_string()
                 }
                 self.attributes
                     .insert(format!("positional_{}", idx + 1), attribute.to_string());
@@ -258,50 +246,6 @@ impl ElementMetadata {
             }
         }
     }
-}
-
-fn key_values_from_named_attribute(attribute: &str) -> Result<(String, Vec<&str>)> {
-    match RE_NAMED.captures(attribute) {
-        Some(captures) => {
-            let (_, [named, mut values_str]) = captures.extract();
-            // remove quotes values
-            if values_str.starts_with('"') {
-                values_str = &values_str[1..values_str.len() - 1]
-            }
-            Ok((
-                named.to_string(),
-                values_str.split(' ').collect::<Vec<&str>>(),
-            ))
-        }
-        None => Err(anyhow!("Not a named attribute")),
-    }
-}
-
-/// Extracts included page ranges from the "lines=" attribute of an include directive
-pub fn extract_page_ranges(ranges_str: &str) -> Vec<i32> {
-    dbg!(&ranges_str);
-    let mut ranges: Vec<i32> = vec![];
-
-    let mut _ranges = ranges_str.split(";").peekable();
-    while let Some(range) = _ranges.next() {
-        let parts: Vec<&str> = range.split("..").collect();
-        if !parts.is_empty() {
-            let start = parts[0].parse::<i32>().unwrap_or(0);
-            if parts.len() == 2 {
-                // if there is an end value
-                if let Some(mut end) = parts[1].parse::<i32>().ok() {
-                    end += 1; // because we want an inclusive range 
-                    ranges.extend(start..end)
-                } else if _ranges.peek().is_none() {
-                    ranges.push(start);
-                    ranges.push(-1);
-                }
-            } else {
-                ranges.push(start);
-            }
-        }
-    }
-    ranges
 }
 
 #[cfg(test)]
