+++ src/graph/mod.rs
@@ -10,7 +10,6 @@ pub mod asg;
 pub mod blocks;
 pub mod inlines;
 pub mod lists;
-pub mod macros;
 pub mod metadata;
 pub mod nodes;
 pub mod substitutions;
